requirements.txt
 azure-keyvault-secrets==4.4.0
 azure-keyvault-keys==4.5.1